src/requirements.txt
git+https://github.com/usdot-its-jpo-data-portal/ode-output-validator-library@master#egg=odevalidator
pyyaml
pysqs-extended-client
requests
Markdown==3.2.1
